extract functions and cleanup use-ui-manager
